refactor: move worker AI types to windmill-ai crate
